fix mismatch fft

+++ src/Storage.py
@@ -31,18 +31,20 @@ class Storage():
         self.ref_connection_str = ref_connection_str
         self.hyp_connection_str = hyp_connection_str
         self.tmp_dir = tmp_dir
-        self.backup_blob_service_client = BlobServiceClient.from_connection_string(self.hyp_connection_str) 
-        try:
-            backup_container_client = self.backup_blob_service_client.create_container(self.container_name)
-        except:
-            pass
-        self.backup_tmp = f"{tmp_dir}/hypotheses/{self.container_name}"
+        self.backup_blob_service_client = None
+        if(hyp_connection_str != None):
+            self.backup_blob_service_client = BlobServiceClient.from_connection_string(self.hyp_connection_str) 
+            try:
+                backup_container_client = self.backup_blob_service_client.create_container(self.container_name)
+            except:
+                pass
+            self.backup_tmp = f"{tmp_dir}/hypotheses/{self.container_name}"
 
-        if(not os.path.exists(self.backup_tmp)):
-            os.makedirs(self.backup_tmp)
-        else:
-            shutil.rmtree(self.backup_tmp)
-            os.makedirs(self.backup_tmp)
+            if(not os.path.exists(self.backup_tmp)):
+                os.makedirs(self.backup_tmp)
+            else:
+                shutil.rmtree(self.backup_tmp)
+                os.makedirs(self.backup_tmp)
 
     def download_references(self):
         # Create the BlobServiceClient object which will be used to create a container client
@@ -65,19 +67,20 @@ class Storage():
         
 
     def backup(self, result):
-        # Create a file in local data directory to upload and download
-        # hyp_ref_score_timestamp_
-        time = result['time_end_record'].replace(" ", "T")
-        dists = '_'.join(["{:.3f}".format(x) for x in result['dists']])
-        local_file_name = f"hyp_{time}_{dists}.wav"
-        upload_file_path = os.path.join(self.backup_tmp, local_file_name)
-        print(upload_file_path)
-        write(upload_file_path, result['sampling_rate'], np.asarray(result['data']))
-        # Create a blob client using the local file name as the name for the blob
-        blob_client = self.backup_blob_service_client.get_blob_client(container=self.container_name, blob=local_file_name)
-        # Upload the created file
-        with open(upload_file_path, "rb") as data:
-            blob_client.upload_blob(data)
-        
-        os.remove(upload_file_path) 
+        if(self.backup_blob_service_client != None):
+            # Create a file in local data directory to upload and download
+            # hyp_ref_score_timestamp_
+            time = result['time_end_record'].replace(" ", "T")
+            dists = '_'.join(["{:.3f}".format(x) for x in result['dists']])
+            local_file_name = f"hyp_{time}_{dists}.wav"
+            upload_file_path = os.path.join(self.backup_tmp, local_file_name)
+            print(upload_file_path)
+            write(upload_file_path, result['sampling_rate'], np.asarray(result['data']))
+            # Create a blob client using the local file name as the name for the blob
+            blob_client = self.backup_blob_service_client.get_blob_client(container=self.container_name, blob=local_file_name)
+            # Upload the created file
+            with open(upload_file_path, "rb") as data:
+                blob_client.upload_blob(data)
+            
+            os.remove(upload_file_path)
